+++ packages/preview-middleware/src/base/config.ts
@@ -76,6 +76,7 @@ export interface TemplateConfig {
     };
     features?: { feature: string; isEnabled: boolean }[];
     locateReuseLibsScript?: boolean;
+    enhancedHomePage?: boolean;
 }
 
 /**
@@ -172,7 +173,8 @@ export function getFlpConfigWithDefaults(config: Partial<FlpConfig> = {}): FlpCo
         apps: config.apps ?? [],
         libs: config.libs,
         theme: config.theme,
-        init: config.init
+        init: config.init,
+        enhancedHomePage: config.enhancedHomePage === true
     } satisfies FlpConfig;
     if (!flpConfig.path.startsWith('/')) {
         flpConfig.path = `/${flpConfig.path}`;
@@ -230,7 +232,7 @@ export function sanitizeRtaConfig(deprecatedRtaConfig: MiddlewareConfig['rta'], 
  *          and its options, such as the layers it applies to and its service URL, if applicable.
  */
 function getFlexSettings(): TemplateConfig['ui5']['flex'] {
-    const localConnectorPath = 'custom.connectors.WorkspaceConnector';
+    const localConnectorPath = 'open/ux/preview/client/flp/WorkspaceConnector';
 
     return [
         { connector: 'LrepConnector', layers: [], url: '/sap/bc/lrep' },
@@ -359,7 +361,8 @@ export function createFlpTemplateConfig(
             },
             bootstrapOptions: ''
         },
-        locateReuseLibsScript: config.libs
+        locateReuseLibsScript: config.libs,
+        enhancedHomePage: config.enhancedHomePage
     } satisfies TemplateConfig;
 }