+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.0.79
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.19
+
+## 0.0.78
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.18
+
+## 0.0.77
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.17
+
+## 0.0.76
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.16
+
+## 0.0.75
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.15
+
 ## 0.0.74
 
 ### Patch Changes