+++ packages/cds-odata-annotation-converter/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/cds-odata-annotation-converter
 
+## 0.5.3
+
+### Patch Changes
+
+-   d873cbd: chore - Add repository metadata in package.json
+-   Updated dependencies [d873cbd]
+    -   @sap-ux/cds-annotation-parser@0.2.8
+
+## 0.5.2
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/odata-vocabularies@0.4.7
+    -   @sap-ux/cds-annotation-parser@0.2.7
+
 ## 0.5.1
 
 ### Patch Changes