+++ packages/create/CHANGELOG.md
@@ -1,5 +1,154 @@
 # @sap-ux/create
 
+## 0.12.28
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.13.18
+-   @sap-ux/flp-config-inquirer@0.2.63
+-   @sap-ux/preview-middleware@0.19.1
+-   @sap-ux/app-config-writer@0.5.43
+
+## 0.12.27
+
+### Patch Changes
+
+-   Updated dependencies [6e32009]
+    -   @sap-ux/preview-middleware@0.19.0
+    -   @sap-ux/app-config-writer@0.5.43
+
+## 0.12.26
+
+### Patch Changes
+
+-   1bca2d8: Adapt Deploy Config Inquirer and Deploy Config Sub Generator to support ADP Projects
+-   Updated dependencies [1bca2d8]
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.34
+
+## 0.12.25
+
+### Patch Changes
+
+-   Updated dependencies [02874f7]
+    -   @sap-ux/preview-middleware@0.18.23
+    -   @sap-ux/app-config-writer@0.5.43
+
+## 0.12.24
+
+### Patch Changes
+
+-   Updated dependencies [e754cb0]
+    -   @sap-ux/adp-tooling@0.13.17
+    -   @sap-ux/preview-middleware@0.18.22
+    -   @sap-ux/flp-config-inquirer@0.2.62
+    -   @sap-ux/app-config-writer@0.5.43
+
+## 0.12.23
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/app-config-writer@0.5.43
+    -   @sap-ux/project-access@1.29.18
+    -   @sap-ux/abap-deploy-config-writer@0.0.100
+    -   @sap-ux/adp-tooling@0.13.16
+    -   @sap-ux/cap-config-writer@0.9.22
+    -   @sap-ux/cards-editor-config-writer@0.4.13
+    -   @sap-ux/flp-config-inquirer@0.2.61
+    -   @sap-ux/mockserver-config-writer@0.8.11
+    -   @sap-ux/odata-service-writer@0.26.13
+    -   @sap-ux/preview-middleware@0.18.21
+    -   @sap-ux/system-access@0.5.33
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.33
+
+## 0.12.22
+
+### Patch Changes
+
+-   Updated dependencies [224494c]
+    -   @sap-ux/odata-service-writer@0.26.12
+    -   @sap-ux/ui5-config@0.26.5
+    -   @sap-ux/adp-tooling@0.13.15
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.32
+    -   @sap-ux/abap-deploy-config-writer@0.0.99
+    -   @sap-ux/app-config-writer@0.5.42
+    -   @sap-ux/cards-editor-config-writer@0.4.13
+    -   @sap-ux/mockserver-config-writer@0.8.10
+    -   @sap-ux/project-access@1.29.17
+    -   @sap-ux/flp-config-inquirer@0.2.60
+    -   @sap-ux/preview-middleware@0.18.20
+    -   @sap-ux/cap-config-writer@0.9.21
+    -   @sap-ux/system-access@0.5.33
+
+## 0.12.21
+
+### Patch Changes
+
+-   Updated dependencies [b012c01]
+    -   @sap-ux/preview-middleware@0.18.19
+    -   @sap-ux/app-config-writer@0.5.41
+
+## 0.12.20
+
+### Patch Changes
+
+-   Updated dependencies [1fd8b3f]
+    -   @sap-ux/preview-middleware@0.18.18
+    -   @sap-ux/app-config-writer@0.5.41
+
+## 0.12.19
+
+### Patch Changes
+
+-   Updated dependencies [f659540]
+    -   @sap-ux/preview-middleware@0.18.17
+    -   @sap-ux/app-config-writer@0.5.41
+
+## 0.12.18
+
+### Patch Changes
+
+-   Updated dependencies [32dafd7]
+    -   @sap-ux/preview-middleware@0.18.16
+    -   @sap-ux/app-config-writer@0.5.41
+
+## 0.12.17
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/app-config-writer@0.5.41
+    -   @sap-ux/ui5-config@0.26.4
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.31
+    -   @sap-ux/adp-tooling@0.13.14
+    -   @sap-ux/preview-middleware@0.18.15
+    -   @sap-ux/system-access@0.5.33
+    -   @sap-ux/flp-config-inquirer@0.2.59
+    -   @sap-ux/abap-deploy-config-writer@0.0.98
+    -   @sap-ux/cards-editor-config-writer@0.4.12
+    -   @sap-ux/mockserver-config-writer@0.8.9
+    -   @sap-ux/odata-service-writer@0.26.11
+    -   @sap-ux/project-access@1.29.16
+    -   @sap-ux/cap-config-writer@0.9.20
+
+## 0.12.16
+
+### Patch Changes
+
+-   Updated dependencies [1ce7fe9]
+    -   @sap-ux/project-access@1.29.15
+    -   @sap-ux/abap-deploy-config-writer@0.0.97
+    -   @sap-ux/adp-tooling@0.13.13
+    -   @sap-ux/app-config-writer@0.5.40
+    -   @sap-ux/cap-config-writer@0.9.19
+    -   @sap-ux/cards-editor-config-writer@0.4.11
+    -   @sap-ux/flp-config-inquirer@0.2.58
+    -   @sap-ux/mockserver-config-writer@0.8.8
+    -   @sap-ux/odata-service-writer@0.26.10
+    -   @sap-ux/preview-middleware@0.18.14
+    -   @sap-ux/system-access@0.5.32
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.30
+
 ## 0.12.15
 
 ### Patch Changes