+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.33.19
+
+### Patch Changes
+
+-   Updated dependencies [d873cbd]
+    -   @sap-ux/fiori-annotation-api@0.4.22
+
+## 0.33.18
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/project-access@1.29.18
+    -   @sap-ux/fiori-annotation-api@0.4.21
+
+## 0.33.17
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.17
+-   @sap-ux/fiori-annotation-api@0.4.20
+
+## 0.33.16
+
+### Patch Changes
+
+-   @sap-ux/fiori-annotation-api@0.4.19
+-   @sap-ux/project-access@1.29.16
+
+## 0.33.15
+
+### Patch Changes
+
+-   Updated dependencies [1ce7fe9]
+    -   @sap-ux/project-access@1.29.15
+    -   @sap-ux/fiori-annotation-api@0.4.18
+
 ## 0.33.14
 
 ### Patch Changes