+++ packages/fiori-generator-shared/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sap-ux/fiori-generator-shared
 
+## 0.9.11
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/project-access@1.29.18
+    -   @sap-ux/telemetry@0.5.72
+
+## 0.9.10
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.17
+-   @sap-ux/telemetry@0.5.71
+
+## 0.9.9
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/btp-utils@1.0.2
+    -   @sap-ux/telemetry@0.5.70
+    -   @sap-ux/project-access@1.29.16
+
+## 0.9.8
+
+### Patch Changes
+
+-   Updated dependencies [1ce7fe9]
+    -   @sap-ux/project-access@1.29.15
+    -   @sap-ux/telemetry@0.5.69
+
 ## 0.9.7
 
 ### Patch Changes