+++ packages/ui5-application-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/ui5-application-writer
 
+## 1.3.3
+
+### Patch Changes
+
+-   Updated dependencies [224494c]
+    -   @sap-ux/ui5-config@0.26.5
+
+## 1.3.2
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/ui5-config@0.26.4
+
 ## 1.3.1
 
 ### Patch Changes