+++ packages/ui5-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui5-config
 
+## 0.26.5
+
+### Patch Changes
+
+-   224494c: Improved add, delete and get methods for service backends.
+
+## 0.26.4
+
+### Patch Changes
+
+-   011c8c5: fix(deps): update dependency axios to v1.8.2 [security]
+
 ## 0.26.3
 
 ### Patch Changes